+++ trakka/components/project/__init__.py
@@ -3,7 +3,7 @@ import click
 
 from trakka.utils.output import table_format_option
 from trakka.utils.output import object_format_option
-from trakka.utils.cmd_filter import hide_admin_cmds
+from trakka.utils.cmd_filter import hide_admin_cmds, show_admin_cmds
 from trakka.utils.options import opt_abbrev, \
     opt_label, \
     opt_view_type, opt_project_client_type, opt_merge_algorithm
@@ -13,6 +13,7 @@ from trakka.utils.options import opt_dashboard_name
 from trakka.utils.options import opt_description
 from trakka.utils.options import opt_organisation
 from trakka.utils.options import opt_status
+from trakka.utils.options import opt_watermark_trees
 from trakka.utils.privilege import PROJECT_RESOURCE
 from trakka.components.log import log_subcommands
 from trakka.components.iam.privilege import privilege_subcommands
@@ -25,6 +26,7 @@ from .dataset import dataset
 from .field import field
 from .metadata import metadata
 from .document import document
+from .organisation import organisation
 
 
 @click.group()
@@ -40,7 +42,7 @@ project.add_command(dataset)
 project.add_command(document)
 project.add_command(log_subcommands(PROJECT_RESOURCE))
 project.add_command(privilege_subcommands(PROJECT_RESOURCE))
-
+project.add_command(organisation, name="org")
 
 @project.command(
         'add',
@@ -53,8 +55,9 @@ project.add_command(privilege_subcommands(PROJECT_RESOURCE))
 @opt_organisation(help="Requesting organisation abbreviation", required=False)
 @opt_dashboard_name(required=False)
 @opt_label(required=False)
-@opt_project_client_type()
+@opt_project_client_type(required=False, default='aardvark')
 @opt_merge_algorithm()
+@opt_watermark_trees()
 def project_add(
         abbrev: str,
         name: str,
@@ -63,7 +66,8 @@ def project_add(
         dashboard_name: str,
         project_label: str,
         client_type: str,
-        merge_algo: str):
+        merge_algo: str,
+        watermark_trees: bool):
     add_project(abbrev,
                 name,
                 description,
@@ -71,7 +75,8 @@ def project_add(
                 dashboard_name,
                 project_label,
                 client_type,
-                merge_algo)
+                merge_algo,
+                watermark_trees)
 
 
 @project.command(
@@ -88,6 +93,7 @@ def project_add(
 @opt_project_client_type(required=False)
 @opt_merge_algorithm(required=False)
 @opt_status(help="Set current project status")
+@opt_watermark_trees(default=None)
 def project_update(
         project_abbrev: str,
         name: str,
@@ -97,7 +103,8 @@ def project_update(
         project_label: str,
         client_type: str,
         merge_algo: str,
-        status: str
+        status: str,
+        watermark_trees: bool
 ):
     update_project(project_abbrev,
                    name,
@@ -107,7 +114,8 @@ def project_update(
                    project_label,
                    client_type,
                    merge_algo,
-                   status)
+                   status,
+                   watermark_trees)
 
 
 @project.command('set-dashboard', hidden=hide_admin_cmds())