+++ trakka/components/project/funcs.py
@@ -45,7 +45,9 @@ def add_project(
         dashboard_name: str,
         project_label: str,
         client_type: str,
-        merge_algo: str):
+        merge_algo: str,
+        watermark_trees: bool
+):
     return api_post(
         path=PROJECT_PATH,
         data={
@@ -58,6 +60,7 @@ def add_project(
             "label": project_label,
             "clientType": client_type,
             "mergeAlgorithm": merge_algo, 
+            "watermarkTrees": watermark_trees
        }
     )
 
@@ -73,6 +76,7 @@ def update_project(
         client_type: str,
         merge_algorithm: str,
         status: str,
+        watermark_trees: bool
 ):
     project = get_project_by_abbrev(project_abbreviation)
     
@@ -87,6 +91,7 @@ def update_project(
             'clientType',
             'mergeAlgorithm',
             'status',
+            'watermarkTrees'
         ]},
     }
     
@@ -109,6 +114,8 @@ def update_project(
         put_project['mergeAlgorithm'] = merge_algorithm
     if status is not None:
         put_project['status'] = status
+    if watermark_trees is not None:
+        put_project['watermarkTrees'] = watermark_trees
         
     return api_put(
         path=f"{PROJECT_PATH}/{project_abbreviation}",